Put setup test session code in API utils for tests
